将JWebTop公共的一些类抽出来，并且加入FastIPC（https://github.com/washheart/fastipc）相关代码。

+++ JWebTop_c/common/fastipc/Client.cpp
@@ -56,7 +56,7 @@ namespace fastipc{
 		if (!memBuf)return ERR_ClientCreate;
 		if (len <= blockSize)return writeBlock(userMsgType, userValue, userShortStr, pBuff, len, NULL, MSG_TYPE_NORMAL); // 可以一次性写完
 		DWORD idx = 0, tmp = len%blockSize;
-		DWORD result = -1;
+		DWORD result = (DWORD)-1;
 		len = len - tmp;
 		char * id = jw::GenerateGuid();
 		len = len - blockSize;// 多减一次，避免在while循环内判断是否是最后的数据包
